\name{comparison}
\alias{comparison}
\title{Compares the output of \link{estimate.rds} for different \eqn{\theta} initialization values.}
\usage{
  comparison(estimation)
}
\arguments{
  \item{estimation}{An estimate.rds  object}
}
\value{
  Used only for side effects.
}
\description{
	Compares the output of \link{estimate.rds} for different \eqn{\theta} initialization values.
}
\author{
  Jonathan D. Rosenblatt \email{john.ros.work@gmail.com}
}
\seealso{
  \code{\link{estimate.rds}}  
}

