\name{ranks.sum}
\alias{ranks.sum}
\title{Utility functions for creating snowball matrix.}
\usage{
  ranks.sum(data)
}
\arguments{
  \item{data}{ An integer vector of the degrees of each individual sampled.}
}
\value{
  The sum of degrees for each rank
}
\description{
  Utility functions for creating snowball matrix.
}
\author{
  Jonathan D. Rosenblatt \email{john.ros.work@gmail.com}
}

