\docType{package}
\name{chords-package}
\alias{chords}
\alias{chords-package}
\title{Population size estimation for respondent driven sampling}
\description{
  Estimates population size and degree distribution in
  respondent driven samples(RDS).
}
\details{
  \tabular{ll}{ Package: \tab chords\cr Type: \tab Package\cr
  Version: \tab 0.5\cr Date: \tab 2011-12-25\cr License:
  \tab GPL (>= 2)\cr }

  Maximum likelihood estimation of population size using
  the methodology in the reference. See \link{estimate.rds} for details.
}
\author{
  Jonathan D. Rosenblatt \email{john.ros@gmail.com}
}
\references{
	Respondent Driven Sampling as an Epidemic Process
	Berchenko Y., Rosenblatt J.D., White R.G.,Frost S.D.W. (Submitted)    
}

