\name{estimate.rds}
\alias{estimate.rds}
\alias{var.theta}


\title{ Degree distribution estimation in respondent driven samples }
\usage{
  	estimate.rds(sampled.degree.vector, Sij, initial.thetas, beta.grid.length, arc=FALSE, control=generate.rds.control())
  	     
    var.theta(sampled.degree.vector, Sij, Njs, theta, beta)                 
    
    %%estimate.rds.two.stage(sampled.degree.vector, Sij, initial.values, arc=FALSE, control=generate.rds.control(), all.solutions=FALSE)
    
    %%estimate.rds.grid(sampled.degree.vector, Sij , grid.values , control=generate.rds.control(maxit = 10, method = "Nelder-Mead"))
}


\arguments{
  \item{sampled.degree.vector}{An integer vector of the degree of each individual sampled (including zeroes).}

  \item{Sij}{An integer matrix of the counts of individuals sampled with rank i up to sampling period j. Row names are required and represent degrees.}
  
  \item{initial.thetas}{Numeric vector of initial \eqn{\theta} values to search over.}
  
  \item{beta.grid.length}{The number of \eqn{\beta} values to try for each \eqn{\theta} value.}
  
   \item{arc}{Deprecated. }
   
  \item{control}{List of control parameters. Generated by \code{\link{generate.rds.control}}. }
  
  \item{theta}{ Numeric value of \eqn{\theta}.  }
  
  \item{Njs}{ Numeric vector of fhe frequecies of degree \eqn{j}. }
  
  \item{beta}{ Numeric value of \eqn{\beta} }
}


\description{
    ML estimation of population size and degree distribution in respondent driven samples. 
}


\value{
\code{estimate.rds} returns a list with the estimates for each starting values. 
Each components includes:
	\item{beta}{See in \emph{details}}
	\item{theta}{See in \emph{details}}
	\item{Nj}{See in \emph{details}}
	\item{likelihood.optimum}{The maximal (log) likelihood.}
	\item{call}{The function call.}	
  
\code{var.theta} returns an estimate of the asymptotic variance of \eqn{\theta}.

\code{generate.rds.control} returns a list of control values needed by \code{estimate.rds}. At present this includes only the number of \code{\link{optim}} iterations.
}


\details{
	\code{estimate.rds} performs maximum likelihood estimation of the population
  size, degree distribution and "coefficient of discoverability" as described in the reference.   
  
   The method assumes the probability of sampling an individual with degree i at time j is given by: 
  \deqn{\beta I[j] (N[i]-S[i,j]) i^\theta }
  where   \eqn{\beta} is the base rate of sampling, \eqn{\theta} is the "coefficient of discoverability", \eqn{I[j]} it the number of individuals sampled up to time \eqn{j}, 
   \eqn{N[i]} is the number of individuals with degree \eqn{i} in the population, and \eqn{S[i,j]} is the number of individuals with degree \eqn{i} sampled up to time \eqn{j}.
   
   The function \code{estimate.rds} has to be given some initialization values. By default, it is given a grid of \eqn{\theta} values, but it can also be given a list including other parameter values (as in the example).           
   }
   
   
\examples{
%\dontrun{
%#### Commented due to long run time ####

#### Using simulated data:
Njs<- c(100,1000,200,200)
names(Njs)<- c("10","20","100","1000")
Njs<- as.table(Njs)
theta<- 1.5
beta<- 3e-9
degree.sampled.vec<- generate.sample(theta, Njs, beta, sample.length=1e3)
#(.Nj.table <- table(degree.sampled.vec))
matplot(t(.Sij <- make.Sij(degree.sampled.vec)), type="s")
plot(.S <- compute.S(.Sij), type="s")
estimate.rds(degree.sampled.vec, .Sij, initial.thetas=seq(-2,2,length=20), beta.grid.length=20)
%}
}

\author{
  Jonathan D. Rosenblatt \email{john.ros.work@gmail.com}
}


\references{
	Respondent Driven Sampling as an Epidemic Process
	Berchenko Y., Rosenblatt J.D., White R.G.,Frost S.D.W. (Submitted) 
}


