\name{generate.rds.control}

\alias{generate.rds.control}

\title{Generator for parameters controlling \code{\link{estimate.rds}}}
\description{
Generator for parameters controlling \code{\link{estimate.rds}}. Used in past version. Presently of no use.
}
\usage{
generate.rds.control()
}

\arguments{ }

\value{
Wraps its arguments as a named list needed for \code{\link{estimate.rds}}.
}
\author{
Jonathan Rosenblatt<john.ros.work@gmail.com>
}

\examples{
generate.rds.control()
}
