\docType{package}
\name{chords-package}
\alias{chords}
\alias{chords-package}
\title{Population size estimation for respondent driven sampling}
\description{
  Estimates population size and degree distribution in
  respondent driven samples (RDS).
}
\details{
  Maximum likelihood estimation of population size using the methodology in the reference. 
  The fundamental idea is of modeling the sampling as an epidemic process.
  See \code{\link{estimate.b.k}} for details.
}
\author{
  Jonathan D. Rosenblatt \email{john.ros.work@gmail.com}
}
\references{
	[1] Berchenko, Yakir, Jonathan Rosenblatt, and Simon D. W. Frost. "Modeling and Analyzing Respondent Driven Sampling as a Counting Process."" arXiv:1304.3505, April 11, 2013. HTTP://arXiv.org/abs/1304.3505.
}

